requirements.txt
flask
flask-jwt-extended
flasgger
psycopg2-binary
python-dotenv